requirements.txt
dhlab==2.2.1
pandas==1.0.1
streamlit
matplotlib
numpy
requests==2.23.0
Pillow==7.0.0
urllib3==1.25.8
